Add add player modal - refactor add player into a modal - add sharp dependency - update test snapshot - clean up logic in add player modal
