+++ package.json
@@ -22,7 +22,8 @@
     "next": "^14.2.23",
     "react": "^18",
     "react-beautiful-dnd": "^13.1.1",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "sharp": "^0.33.5"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.6.2",
